backend/requirements.txt
-i https://pypi.org/simple
anyio==3.7.1; python_version >= '3.7'
astroid==3.0.1; python_full_version >= '3.8.0'
boto3==1.28.78
botocore==1.31.78; python_version >= '3.7'
cachetools==5.3.2
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dill==0.3.7; python_version >= '3.11'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.104.1
gdal==3.6.4
geoalchemy2==0.14.2
greenlet==3.0.1; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
isort==5.12.0; python_full_version >= '3.8.0'
jmespath==1.0.1; python_version >= '3.7'
mangum==0.17.0
mccabe==0.7.0; python_version >= '3.6'
numpy==1.26.1; python_version < '3.13' and python_version >= '3.9'
packaging==23.2; python_version >= '3.7'
pillow==10.1.0
platformdirs==3.11.0; python_version >= '3.7'
psycopg2-binary==2.9.9
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==1.10.13
pylint==3.0.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9; python_version >= '3.6' and python_version < '4'
s3transfer==0.7.0; python_version >= '3.7'
shapely==2.0.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==1.4.50; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy2-stubs==0.0.2a36; python_version >= '3.6'
sqlmodel==0.0.11
starlette==0.27.0; python_version >= '3.7'
tomlkit==0.12.2; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.7'
uvicorn==0.24.0
